[Refactor] 새 비밀번호 생성 페이지 폴더 위치 변경 후 리펙터링 #403
