+++ app/src/pages/login/findPassword/setNewPassword/Presenter.tsx
@@ -8,91 +8,92 @@ interface Props {
     confirmChangePassword: string;
   };
   showHidenPassword: { [key: string]: boolean };
-  popupInfo: { view: boolean; message: string };
-  handleHidenPassword: (str: string) => void;
-  handleOnChange: (e: React.ChangeEvent<HTMLInputElement>, key: string) => void;
-  submitAble: () => boolean;
+  handleHidenPassword: (key: string) => void;
+  handleOnChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
   onSubmit: (
     e: React.FormEvent<HTMLFormElement> | React.FormEvent<HTMLButtonElement>,
   ) => void;
-  closePopup: () => void;
 }
 
 function Presenter(props: Props) {
   const {
     inputValue,
     showHidenPassword,
-    popupInfo,
     handleHidenPassword,
     handleOnChange,
-    submitAble,
     onSubmit,
-    closePopup,
   } = props;
+  const { email, changePassword, confirmChangePassword } = inputValue;
+
+  const toggleImgSrc = (key: string) => {
+    return showHidenPassword[key]
+      ? '/img/hide-password.png'
+      : '/img/show-password.png';
+  };
+
+  const submitAble = () => {
+    if (email.length && changePassword.length && confirmChangePassword.length)
+      return true;
+    else return false;
+  };
 
   return (
-    <>
-      <Wrapper>
-        <Container onSubmit={onSubmit}>
-          <section>
-            <p>새 비밀번호 설정</p>
-            <p>새로운 비밀번호를 설정해 주세요.</p>
-            <Line />
-          </section>
-          <section>
-            <label className="title-label">이메일</label>
-            <Input
-              type={'email'}
-              value={inputValue.email}
-              placeholder="이메일을 입력해 주세요."
-              onChange={e => handleOnChange(e, 'email')}
-            />
-          </section>
-          <section>
-            <label className="title-label">새 비밀번호</label>
-            <div>
-              <ShadowLabel>
-                <Input
-                  type={showHidenPassword.new ? 'text' : 'password'}
-                  password
-                  value={inputValue.changePassword}
-                  placeholder="새로운 비밀번호를 입력해 주세요. (8~15자)"
-                  onChange={e => handleOnChange(e, 'changePassword')}
-                />
-                <img
-                  src={
-                    showHidenPassword.new
-                      ? '/img/hide-password.png'
-                      : '/img/show-password.png'
-                  }
-                  onClick={() => handleHidenPassword('new')}
-                />
-              </ShadowLabel>
-              <ShadowLabel>
-                <Input
-                  type={showHidenPassword.check ? 'text' : 'password'}
-                  password
-                  value={inputValue.confirmChangePassword}
-                  placeholder="새로운 비밀번호를 다시 한 번 입력해 주세요."
-                  onChange={e => handleOnChange(e, 'confirmChangePassword')}
-                />
-                <img
-                  src={
-                    showHidenPassword.check
-                      ? '/img/hide-password.png'
-                      : '/img/show-password.png'
-                  }
-                  onClick={() => handleHidenPassword('check')}
-                />
-              </ShadowLabel>
-            </div>
-          </section>
-          <SubmitButton able={submitAble()} onSubmit={onSubmit}>
-            변경하기
-          </SubmitButton>
-        </Container>
-      </Wrapper>
-    </>
+    <Wrapper>
+      <Container onSubmit={onSubmit}>
+        <section>
+          <p>새 비밀번호 설정</p>
+          <p>새로운 비밀번호를 설정해 주세요.</p>
+          <Line />
+        </section>
+        <section>
+          <label className="title-label">이메일</label>
+          <Input
+            name="email"
+            type="email"
+            value={inputValue.email}
+            placeholder="이메일을 입력해 주세요."
+            onChange={handleOnChange}
+          />
+        </section>
+        <section>
+          <label className="title-label">새 비밀번호</label>
+          <div>
+            <ShadowLabel>
+              <Input
+                password
+                name="changePassword"
+                type={showHidenPassword.new ? 'text' : 'password'}
+                value={inputValue.changePassword}
+                placeholder="새로운 비밀번호를 입력해 주세요. (8~15자)"
+                onChange={handleOnChange}
+              />
+              <img
+                id="new"
+                src={toggleImgSrc('new')}
+                onClick={() => handleHidenPassword('new')}
+              />
+            </ShadowLabel>
+            <ShadowLabel>
+              <Input
+                password
+                name="confirmChangePassword"
+                type={showHidenPassword.check ? 'text' : 'password'}
+                value={inputValue.confirmChangePassword}
+                placeholder="새로운 비밀번호를 다시 한 번 입력해 주세요."
+                onChange={handleOnChange}
+              />
+              <img
+                src={toggleImgSrc('check')}
+                onClick={() => handleHidenPassword('check')}
+              />
+            </ShadowLabel>
+          </div>
+        </section>
+        <SubmitButton able={submitAble()} onSubmit={onSubmit}>
+          변경하기
+        </SubmitButton>
+      </Container>
+    </Wrapper>
   );
 }
